+++ CHANGELOG.md
@@ -5,12 +5,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
-- 2020-02-25 : (PLB2B-1045)  : Added  `notConnected` error case for scenario where a service grant is requested in not connected state
-- 2020-02-25 : (PLB2B-1024)  : Added  bluetooth state events to tracking
-
 ### Changed
 ### Fixed
 ### Removed
+
+
+## [3.10.0] - 2020-02-25
+
+### Added
+- 2020-02-25 : (PLB2B-1045)  : Added  `notConnected` error case for scenario where a service grant is requested in not connected state
+- 2020-02-25 : (PLB2B-1024)  : Added  bluetooth state events to tracking
 
 ## [3.9.0] - 2020-02-17